+++ package.json
@@ -34,6 +34,7 @@
     "since 2017-06"
   ],
   "dependencies": {
+    "@reach/router": "^1.2.1",
     "axios": "^0.18.0",
     "dotenv": "^6.2.0",
     "formik": "^1.5.1",
